introduce slices
