--- src/features/todo/App/index.tsx
@@ -1,57 +1,0 @@
-import * as React from 'react';
-import * as TodoActions from '../actions/todos';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
-import { RouteComponentProps } from 'react-router';
-import { RootState } from '../../../store';
-import MainSection from '../components/MainSection';
-import Header from '../components/Header';
-import { Container, Row, Col } from "react-bootstrap";
-
-export namespace App {
-  export interface Props extends RouteComponentProps<void> {
-    todos: TodoItemData[];
-    actions: typeof TodoActions;
-  }
-
-  export interface State {
-    /* empty */
-  }
-}
-
-class App extends React.Component<App.Props, App.State> {
-
-  constructor(props, state) {
-    super(props, state)
-  }
-
-  render() {
-    const { todos, actions, children } = this.props;
-    console.log('>>> '+JSON.stringify(this.props));
-    return (
-      <Container fluid>
-        <Row className="mt-5 justify-content-md-center" >
-          <Col className="col-md-4">
-          <Header addTodo={actions.addTodo} />
-          <MainSection todos={todos} actions={actions} />
-          {children}
-          </Col>
-        </Row>
-      </Container>
-    );
-  }
-}
-
-function mapStateToProps(state: RootState) {
-  return {
-    todos: state.todos
-  };
-}
-
-function mapDispatchToProps(dispatch) {
-  return {
-    actions: bindActionCreators(TodoActions as any, dispatch)
-  };
-}
-
-export default (connect(  mapStateToProps,  mapDispatchToProps) as any) (App)
